+++ api/src/main/java/fr/kksdev/budget/api/enums/EntityType.java
@@ -3,5 +3,6 @@ package fr.kksdev.budget.api.enums;
 public enum EntityType {
     SUBSCRIPTION,
     DEBT,
-    BUDGET
+    BUDGET,
+    TRANSACTION
 }